refactor: split route info tool helpers
